+++ apps/src/lab2/views/components/editor/code-editor.module.scss
@@ -1,12 +1,3 @@
 .codeEditorContainer {
   height: 100%;
-
-  .aiTutorTooltip {
-    border-radius: 6px;
-    border: 1px solid var(--borders-neutral-primary);
-    background: var(--background-neutral-primary);
-
-    /* Toast Shadow */
-    box-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.07), 0 4px 7px 0 rgba(0, 0, 0, 0.07);
-  }
 }
